Binding args using lambdas instead of variadic templates
This is both prettier and easier to understand. I'm tempted to throw away everything I've done with the variadic templates and force the user to use a zero-argument lambda all the time. (Possibly even omit the return value, which would take care of another rough edge!) The only reason I hesitate is because that would amount to taking a function which is possibly pure (though possibly not), and then make it as impure as possible, which feels like the opposite of my overall goal of increasing purity. On the other hand, I wouldn't have to stress about horrible C++ reference forwarding semantics all the time...