Re-add USE_CUDA CMake flag
We still need the ability to compile the rest of PHASM even without CUDA. This time we only use the flag at the CMake level, not in the code itself. Eventually we can refactor the examples so that the CUDA dependency lives behind an independent plugin, and the examples + PHASM will work whether CUDA is present or not.