Clean up example macros
CMake was failing because it expected there to be an `exampleB4d.out` file,
which was never committed because of .gitignore. There is no point in having
the `exampleB4d.out` file to begin with, as far as I can tell. A bunch of other
macros included in each example appear to be redundant, so I removed those as
well.